Add better portability between dev environments, and a better camera.
There is now a basic framework to set up systemd and other config files,
based on templates and a config JSON file.  This can be improved, of
course.

For now, the new "getting started" flow is to cd etc, then run
setup_libraries.sh, and then setup_host.sh.  This should work on
distros with systemd, which means Debian Jesse or later.  That
doesn't cover Cloud9, but the preferred way to run servers under Cloud9
is different anyway.

Also, improved camdaemon: with the config option use_video_port=true,
it now can run at 30 fps.  Also, it can now probe the camera, and use
the real one or fake one depending on what's available.  (Previously,
it was a static option.)